+++ Changelog.md
@@ -1,4 +1,5 @@
-# gov-ua-ui changelog from v0.4.3. to v0.4.4.
+# gov-ua-ui changelog from v0.4.4. to v0.4.5.
 
-1. Fixed Datepicker transition from years to days without hide calendar.
+1. Added possibility to set scroll margins when navigate through anchor links.
+2. Fixed Tooltip behavior when parent element scrolling horizontally.